Adapt to most recent version of the JetMatchingP8internal.h code. DJR estimation is now done using slowJet

+++ SimDataFormats/GeneratorProducts/interface/GenEventInfoProduct.h
@@ -40,7 +40,7 @@ class GenEventInfoProduct {
 	const std::vector<double> &binningValues() const { return binningValues_; }
 	bool hasBinningValues() const { return !binningValues_.empty(); }
 
-        const std::vector<float> &DJRValues() const { return DJRValues_; }
+        const std::vector<double> &DJRValues() const { return DJRValues_; }
         bool hasDJRValues() const { return !DJRValues_.empty(); }
 
         const std::vector<int> &nMEPartons() const { return nMEPartons_; }
@@ -62,7 +62,7 @@ class GenEventInfoProduct {
 	void setBinningValues(const std::vector<double> &values)
 	{ binningValues_ = values; }
  
-	void setDJR(const std::vector<float> &values)
+	void setDJR(const std::vector<double> &values)
 	{DJRValues_ = values;}
 
 	void setNMEPartons(const std::vector<int> &values)
@@ -90,7 +90,7 @@ class GenEventInfoProduct {
 	// will contain the information what physical
 	// quantity these values actually belong to
 	std::vector<double>	binningValues_;
-        std::vector<float>      DJRValues_;
+        std::vector<double>      DJRValues_;
         std::vector<int>     	nMEPartons_;
 };
